squash! test/AspNetCore/HttpSysCommunicationListenerTest.cs
Address the following finding.

### ⚠️ Inconsistent port fuzzing (elevated from 💡)
*Reported by: opus. Cross-check: gemini Agree, gpt Agree.*

At [lines 92-102](test/AspNetCore/HttpSysCommunicationListenerTest.cs#L92-L102) the "other" port uses `fuzzy.UInt16()` while the target uses `fuzzy.Int32()`. `Int32` can produce negative or out-of-range port values that don't represent real Service Fabric endpoints; sibling tests use `UInt16` for ports (e.g. [test/AspNetCore/AspNetCoreCommunicationListenerTest.cs](test/AspNetCore/AspNetCoreCommunicationListenerTest.cs#L344)).

**Fix:** Use `fuzzy.UInt16()` for both.

+++ test/AspNetCore/HttpSysCommunicationListenerTest.cs
@@ -109,7 +109,7 @@ public abstract class HttpSysCommunicationListenerTest
                 Name = endpointName,
                 Protocol = protocol,
             };
-            int port = fuzzy.Int32();
+            int port = fuzzy.UInt16();
             endpoint.Property<int>().Set(port);
             context.CodePackageActivationContext.GetEndpoints().Add(endpoint);
